+++ pages/api/sessionClient.ts
@@ -1,7 +1,7 @@
+﻿import {assertTokenIsValid} from "../../helpers/tokenHelpers";
 ﻿import getConfig from 'next/config';
 import fetch from "node-fetch";
 import {GetServerSideProps} from "next";
-import {assertTokenIsValid} from "../helpers/tokenHelpers";
 import {CandidateTestStatus} from "./candidateApiClientModule";
 const {publicRuntimeConfig} = getConfig();
 const baseUrl = publicRuntimeConfig.API_URL;
@@ -22,24 +22,18 @@ export interface NewTestSubmission{
     testId: number;
     testAnswer: string;
 }
+
 export async function getSessionCandidate(token: string | string[] | undefined): Promise<SessionCandidate> {
-   try {
+    const {publicRuntimeConfig} = getConfig();
+    const baseUrl = publicRuntimeConfig.API_URL;
+    try {
         const result = await fetch(
             `${baseUrl}/sessions/${token}`
         );
-        return await result.json();
+        const data = await result.json();
+        return data;
     } catch (error) {
         console.error(error);
         return error.message;
     }
-}
-
-export async function addTestSubmission(tokenId: string, newTestSubmission: NewTestSubmission) {
-    return await fetch(`${baseUrl}/sessions/${tokenId}`, {
-        method: "POST",
-        headers: {
-            "Content-Type": "application/json"
-        },
-        body: JSON.stringify(newTestSubmission),
-    })
 }
